+++ new_bit_array.go
@@ -28,15 +28,15 @@ func NewBitArrayOfLength(length uint64) BitArray {
 }
 
 // Returns an array of size 'bits' initialized with the given 'sparsity'.
-func NewBitArrayOfSparsity(length uint64, sparsity float32) BitArray {
+func NewBitArrayOfSparsity(length uint64, sparsity float64) BitArray {
 	if sparsity < 0 || sparsity > 1 {
 		panic("Sparsity must be between 0 and 1")
 	}
 	if sparsity == 0 {
-		sparsity = float32(rand.Float64())
+		sparsity = float64(rand.Float64())
 	}
 	this := NewBitArrayOfLength(length)
-	space := uint64(float32(length) * sparsity)
+	space := uint64(float64(length) * sparsity)
 	for i := uint64(0); i < space; i++ {
 		this.Write(i, true)
 	}
